+++ src/App.jsx
@@ -1,15 +1,13 @@
 import { Route, Routes } from "react-router-dom";
-import Mypage from "./pages/Mypage/Mypage";
 import Admin from "./pages/Admin/Admin";
+import Home from "./pages/Home/Home";
 
 function App() {
-  const username = "ajvlzla0207"; // 유동적으로 받아올 부분
-
   return (
     <>
       <Routes>
-        <Route path="/admin/*" element={ <Admin /> }/>
-        <Route path={`/${username}`} element={<Mypage username={username}/>} />
+        <Route path="/*" element={ <Home />} />
+        <Route path="/admin/*" element={ <Admin /> } />
       </Routes>
     </>
   )
