+++ src/apis/api/mypage.js
@@ -1,5 +1,5 @@
 import instance from "../utils/instance"
 
 export const getLoanDataRequest = async (data) => {
-    return await instance.get(`/mypage/loan/${data.userName}`, data)
+    return await instance.get(`/mypage/loan/${data.username}`, data)
 }
