+++ src/components/Header/Header.jsx
@@ -1,0 +1,14 @@
+/** @jsxImportSource @emotion/react */
+import * as s from "./style"
+
+function Header(props) {
+    return (
+        <div css={s.headerLayout}>
+            <div css={s.header}>
+                <h1>HEADER</h1>
+            </div>
+        </div>
+    );
+}
+
+export default Header;
