--- src/components/readme
@@ -1,1 +1,0 @@
-컴포넌트 모듈
