+++ src/pages/LoanAndReturn/style.js
@@ -1,0 +1,19 @@
+import { css } from "@emotion/react";
+
+export const loanTable = css`
+    border-collapse: collapse;
+    border: 1px solid black;
+
+    font-size: 14px;
+
+    th, td {
+        border: 1px solid black;
+        padding: 10px; 
+    }
+
+    th {
+        background-color: #bab3b3;
+        text-align: left;
+    }
+`;
+
