+++ src/pages/Mypage/Mypage.jsx
@@ -1,39 +1,17 @@
 /** @jsxImportSource @emotion/react */
-import { useState } from "react";
-import { useQuery } from "react-query";
-import { getLoanDataRequest } from "../../apis/api/mypage";
+import * as s from "./style"
 import { Link } from "react-router-dom";
 import { MENUS } from "../../constants/mypageMenu";
-import * as s from "./style"
+import { Route, Routes } from "react-router-dom";
+import UserInfoModification from "../UserInfoModification/UserInfoModification";
+import WishList from "../WishList/WishList";
+import LoanAndReturn from "../LoanAndReturn/LoanAndReturn";
+import Overdue from "../Overdue/Overdue";
+import MypageMain from "../MypageMain/MypageMain";
 
-function Mypage({ username }) {
-    const [ loanList, setLoanList ] = useState([]);
-
-    const searchLoansQuery = useQuery(
-        ["searchLoansQuery"],
-        async () => await getLoanDataRequest({
-            userName: username
-        }),
-        {
-            retry: 0,
-            refetchOnWindowFocus: false,
-            onSuccess: response => {
-                setLoanList(response.data);
-                console.log(response.data)
-            },
-            onError: error => {
-                console.log(error);
-            }
-        }
-    );
-
+function Mypage() {
     return (
         <>
-            <div css={s.headerLayout}>
-                <div css={s.header}>
-                    <h1>HEADER</h1>
-                </div>
-            </div>
             <div css={s.bodyLayout}>
                 <div css={s.profileContainer}>
                     <div css={s.left}>
@@ -58,41 +36,16 @@ function Mypage({ username }) {
                         </div>
                     </div>
                     <div css={s.right}>
-                        <table css={s.loanTable}>
-                            <thead>
-                                <tr>
-                                    <th>대출번호</th>
-                                    <th>재고번호</th>
-                                    <th>제목</th>
-                                    <th>저자</th>
-                                    <th>출판사</th>
-                                    <th>대출일자</th>
-                                    <th>반납예정일</th>
-                                    <th>반납일자</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {
-                                    loanList.map(
-                                        loan => 
-                                            <tr key={loan.loanId}>
-                                                <td>{loan.loanId}</td>
-                                                <td>{loan.bookStockId}</td>
-                                                <td>{loan.bookName}</td>
-                                                <td>{loan.authorName}</td>
-                                                <td>{loan.publisherName}</td>
-                                                <td>{loan.loanDate}</td>
-                                                <td>{loan.dueDate}</td>
-                                                <td>{loan.returnDate === null ? "대출중" : loan.returnDate}</td>
-                                            </tr>
-                                    )
-                                }
-                            </tbody>
-                        </table>
+                        <Routes>
+                            <Route path="/" element={ <MypageMain />}/>
+                            <Route path="/info" element={ < UserInfoModification/> }/>
+                            <Route path="/wishlist" element={ < WishList/> }/>
+                            <Route path="/loan" element={ < LoanAndReturn/> }/>
+                            <Route path="/overdue" element={ < Overdue/> }/>
+                        </Routes>
                     </div>
                 </div>
             </div>
-            
         </>    
     );
 }
