+++ src/pages/MypageMain/MypageMain.jsx
@@ -1,0 +1,12 @@
+/** @jsxImportSource @emotion/react */
+import * as s from "./style"
+
+function MypageMain(props) {
+    return (
+        <div>
+            <h1>마이페이지 메인</h1>
+        </div>
+    );
+}
+
+export default MypageMain;
