+++ src/pages/MypageMain/style.js
@@ -1,0 +1,2 @@
+import { css } from "@emotion/react";
+
