+++ src/pages/Overdue/Overdue.jsx
@@ -1,9 +1,10 @@
-import React from 'react';
+/** @jsxImportSource @emotion/react */
+// import * as s from "./style"
 
 function Overdue(props) {
     return (
         <div>
-            
+            <h1>연체 목록</h1>
         </div>
     );
 }
