+++ src/pages/UserInfoModification/UserInfoModification.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 function UserInfoModification(props) {
     return (
         <div>
-            
+            <h1>개인정보수정</h1>
         </div>
     );
 }
