+++ src/pages/WishList/WishList.jsx
@@ -1,9 +1,10 @@
-import React from 'react';
+/** @jsxImportSource @emotion/react */
+// import * as s from "./style"
 
 function WishList(props) {
     return (
         <div>
-            
+            <h1>위시리스트</h1>
         </div>
     );
 }
