+++ src/pages/WishList/style.js
@@ -1,0 +1,3 @@
+import { css } from "@emotion/react";
+
+
